Concurrency fixes
