+++ src/main.rs
@@ -35,13 +35,13 @@ mod cli;
 /// Configuration strectures for Eva01
 mod config;
 
-/// Transactio manager
+mod transaction_checker;
 mod transaction_manager;
 
 /// Crossbar client
 mod crossbar;
 
-#[tokio::main(flavor = "multi_thread", worker_threads = 6)]
+#[tokio::main(flavor = "multi_thread", worker_threads = 16)]
 async fn main() -> Result<(), Box<dyn Error>> {
     // Assemble logger, with INFO as default log level
     Builder::from_env(env_logger::Env::default().default_filter_or("info")).init();
