+++ src/transaction_manager.rs
@@ -18,18 +18,15 @@ use solana_sdk::{
     system_instruction::transfer,
     transaction::VersionedTransaction,
 };
-use std::sync::Arc;
-use std::{str::FromStr, thread};
+use std::str::FromStr;
 
 /// Manages transactions for the liquidator and rebalancer
 #[allow(dead_code)]
 pub struct TransactionManager {
-    rx: Receiver<TransactionData>,
-    ack_tx: Sender<Pubkey>,
     keypair: Keypair,
-    rpc: Arc<RpcClient>,
-    jito_sdk: Arc<JitoJsonRpcSDK>,
+    jito_sdk: JitoJsonRpcSDK,
     jito_tip_account: Pubkey,
+    rpc_client: RpcClient,
     lookup_tables: Vec<AddressLookupTableAccount>,
 }
 
@@ -63,20 +60,9 @@ impl RawTransaction {
     }
 }
 
-#[derive(Debug)]
-struct BundleStatus {
-    confirmation_status: Option<String>,
-    err: Option<serde_json::Value>,
-    transactions: Option<Vec<String>>,
-}
-
 impl TransactionManager {
     /// Creates a new transaction manager
-    pub fn new(
-        rx: Receiver<TransactionData>,
-        ack_tx: Sender<Pubkey>,
-        config: GeneralConfig,
-    ) -> anyhow::Result<Self> {
+    pub fn new(config: GeneralConfig) -> anyhow::Result<Self> {
         let keypair = read_keypair_file(&config.keypair_path)
             .map_err(|e| {
                 error!(
@@ -86,15 +72,14 @@ impl TransactionManager {
             })
             .unwrap();
 
-        let rpc = Arc::new(RpcClient::new_with_commitment(
-            config.rpc_url.clone(),
-            CommitmentConfig::confirmed(),
-        ));
+        debug!("Initializing RPC client with URL: {}", config.rpc_url);
+        let rpc_client =
+            RpcClient::new_with_commitment(config.rpc_url.clone(), CommitmentConfig::confirmed());
 
         // Loads the Address Lookup Table's accounts
         let mut lookup_tables = vec![];
         for table_address in &config.address_lookup_tables {
-            let raw_account = rpc.get_account(table_address)?;
+            let raw_account = rpc_client.get_account(table_address)?;
             let address_lookup_table = AddressLookupTable::deserialize(&raw_account.data)?;
             let lookup_table = AddressLookupTableAccount {
                 key: *table_address,
@@ -103,33 +88,29 @@ impl TransactionManager {
             lookup_tables.push(lookup_table);
         }
 
-        let jito_sdk = Arc::new(JitoJsonRpcSDK::new(&config.block_engine_url, None));
+        // Get JITO tip account
+        let jito_block_engine_url = config.block_engine_url;
+        debug!("Initializing JITO SDK with URL: {}", jito_block_engine_url);
+        let jito_sdk: JitoJsonRpcSDK = JitoJsonRpcSDK::new(&jito_block_engine_url, None);
         let random_tip_account = block_on(jito_sdk.get_random_tip_account())?;
         let jito_tip_account = Pubkey::from_str(&random_tip_account)?;
 
         Ok(Self {
-            rx,
-            ack_tx,
             keypair,
-            rpc,
             jito_sdk,
+            rpc_client,
             jito_tip_account,
             lookup_tables,
         })
     }
 
     /// Starts the transaction manager
-    pub fn start(&mut self) {
-        let (jito_tx, jito_rx) = crossbeam::channel::unbounded::<(Pubkey, String)>();
-        let jito_sdk = Arc::clone(&self.jito_sdk);
-        let ack_tx = self.ack_tx.clone();
-        Self::check_bundle_status(jito_rx, jito_sdk, ack_tx);
-
+    pub fn start(&mut self, jito_tx: Sender<(Pubkey, String)>, txn_rx: Receiver<TransactionData>) {
         info!("Starting the Transaction manager loop.");
         while let Ok(TransactionData {
             transactions,
             ack_id,
-        }) = self.rx.recv()
+        }) = txn_rx.recv()
         {
             debug!("Ack ID: {:?}", ack_id);
 
@@ -167,198 +148,6 @@ impl TransactionManager {
         info!("The Transaction manager loop is stopped.");
     }
 
-    fn check_bundle_status(
-        jito_rx: Receiver<(Pubkey, String)>,
-        jito_sdk: Arc<JitoJsonRpcSDK>,
-        ack_tx: Sender<Pubkey>,
-    ) {
-        let max_retries = 10;
-        let retry_delay = std::time::Duration::from_millis(500);
-        while let Ok((ack_id, uuid)) = jito_rx.recv() {
-            for attempt in 1..=max_retries {
-                debug!(
-                    "Checking bundle {} (ack_id: {}) status (attempt {}/{})",
-                    uuid, ack_id, attempt, max_retries
-                );
-
-                let status_response =
-                    block_on(jito_sdk.get_in_flight_bundle_statuses(vec![uuid.to_string()]));
-                if let Err(e) = status_response {
-                    debug!(
-                        "Failed to check bundle {} (ack_id: {}) status: {:?}",
-                        uuid, ack_id, e
-                    );
-                    continue;
-                }
-
-                let status_response = status_response.unwrap();
-                match status_response.get("result") {
-                    Some(result) => {
-                        match result
-                            .get("value")
-                            .and_then(|value| value.as_array())
-                            .and_then(|statuses| statuses.first())
-                            .and_then(|bundle_status| bundle_status.get("status"))
-                            .and_then(|status| status.as_str())
-                        {
-                            Some("Landed") => {
-                                debug!(
-                                    "({}) Bundle landed on-chain. Checking final status...",
-                                    uuid
-                                );
-                                if let Err(e) = Self::check_final_bundle_status(&jito_sdk, &uuid) {
-                                    error!("({}) Final status: {}", uuid, e.to_string());
-                                }
-                                break;
-                            }
-                            Some("Pending") => {
-                                debug!("({}) Bundle is pending. Waiting...", uuid);
-                            }
-                            Some(status) => {
-                                debug!(
-                                    "({}) Unexpected bundle status: {}. Waiting...",
-                                    uuid, status
-                                );
-                            }
-                            None => {
-                                debug!("({}) Unable to parse bundle status. Waiting...", uuid);
-                            }
-                        }
-                    }
-                    None => match status_response.get("error") {
-                        Some(error) => {
-                            debug!("({}) Error checking bundle status: {:?}", uuid, error);
-                        }
-                        None => {
-                            debug!("({}) Unexpected response format. Waiting...", uuid);
-                        }
-                    },
-                }
-
-                if attempt < max_retries {
-                    debug!(
-                        "Sleeping for {} ms before retrying...",
-                        retry_delay.as_millis()
-                    );
-                    thread::sleep(retry_delay);
-                } else {
-                    ERROR_COUNT.inc();
-                    error!(
-                        "Failed to confirm bundle status: uuid = {}, ack_id = {}",
-                        uuid, ack_id
-                    );
-                    break;
-                }
-
-                debug!("Sending ACK for bundle {} (ack_id: {})", uuid, ack_id);
-                ack_tx.send(ack_id).unwrap();
-            }
-        }
-    }
-
-    fn check_final_bundle_status(jito_sdk: &JitoJsonRpcSDK, uuid: &str) -> anyhow::Result<()> {
-        let max_retries = 10;
-        let retry_delay = std::time::Duration::from_millis(500);
-
-        for attempt in 1..=max_retries {
-            debug!(
-                "({}) Checking final bundle status (attempt {}/{})",
-                uuid, attempt, max_retries
-            );
-
-            let status_response = block_on(jito_sdk.get_bundle_statuses(vec![uuid.to_string()]))?;
-            let bundle_status = Self::get_bundle_status(&status_response)?;
-
-            match bundle_status.confirmation_status.as_deref() {
-                Some("confirmed") => {
-                    debug!(
-                        "({}) Bundle confirmed on-chain. Waiting for finalization...",
-                        uuid
-                    );
-                    Self::check_transaction_error(&bundle_status)?;
-                }
-                Some("finalized") => {
-                    debug!("({}) Bundle finalized on-chain successfully!", uuid);
-                    Self::check_transaction_error(&bundle_status)?;
-                    Self::print_transaction_url(&bundle_status);
-                    return Ok(());
-                }
-                Some(status) => {
-                    debug!(
-                        "({}) Unexpected final bundle status: {}. Continuing to poll...",
-                        uuid, status
-                    );
-                }
-                None => {
-                    debug!(
-                        "({}) Unable to parse final bundle status. Continuing to poll...",
-                        uuid
-                    );
-                }
-            }
-
-            if attempt < max_retries {
-                thread::sleep(retry_delay);
-            }
-        }
-
-        Err(anyhow::anyhow!(
-            "({}) Failed to get finalized status after {} attempts",
-            uuid,
-            max_retries
-        ))
-    }
-
-    fn get_bundle_status(status_response: &serde_json::Value) -> anyhow::Result<BundleStatus> {
-        status_response
-            .get("result")
-            .and_then(|result| result.get("value"))
-            .and_then(|value| value.as_array())
-            .and_then(|statuses| statuses.first())
-            .ok_or_else(|| anyhow::anyhow!("Failed to parse bundle status"))
-            .map(|bundle_status| BundleStatus {
-                confirmation_status: bundle_status
-                    .get("confirmation_status")
-                    .and_then(|s| s.as_str())
-                    .map(String::from),
-                err: bundle_status.get("err").cloned(),
-                transactions: bundle_status
-                    .get("transactions")
-                    .and_then(|t| t.as_array())
-                    .map(|arr| {
-                        arr.iter()
-                            .filter_map(|v| v.as_str().map(String::from))
-                            .collect()
-                    }),
-            })
-    }
-
-    fn check_transaction_error(bundle_status: &BundleStatus) -> anyhow::Result<()> {
-        if let Some(err) = &bundle_status.err {
-            if err["Ok"].is_null() {
-                println!("Transaction executed without errors.");
-                Ok(())
-            } else {
-                println!("Transaction encountered an error: {:?}", err);
-                Err(anyhow::anyhow!("Transaction encountered an error"))
-            }
-        } else {
-            Ok(())
-        }
-    }
-
-    fn print_transaction_url(bundle_status: &BundleStatus) {
-        if let Some(transactions) = &bundle_status.transactions {
-            if let Some(tx_id) = transactions.first() {
-                debug!("Transaction URL: https://solscan.io/tx/{}", tx_id);
-            } else {
-                debug!("Unable to extract transaction ID.");
-            }
-        } else {
-            debug!("No transactions found in the bundle status.");
-        }
-    }
-
     /*
 
                if let Some(BundleRejectionError::SimulationFailure(tx_str, msg)) =
@@ -383,7 +172,7 @@ impl TransactionManager {
         &self,
         instructions: BatchTransactions,
     ) -> anyhow::Result<Vec<String>> {
-        let blockhash = self.rpc.get_latest_blockhash()?;
+        let blockhash = self.rpc_client.get_latest_blockhash()?;
 
         let mut txs = Vec::new();
         for raw_transaction in instructions {
