+++ src/utils/mod.rs
@@ -44,7 +44,7 @@ pub struct BatchLoadingConfig {
 
 impl BatchLoadingConfig {
     pub const DEFAULT: Self = Self {
-        max_batch_size: 1000,
+        max_batch_size: 100,
         max_concurrent_calls: 16,
     };
 }
